Promoted IsSentryRequest from Extension method to class method
